+++ src/benchmarking/measure.py
@@ -1,13 +1,11 @@
-import subprocess
-import time
+import subprocess, sys
 
-def main():
+def measure(training_type):
     main_proc = subprocess.Popen(
-        ['python3.10', '-m', 'src.main'],
+        ['python3.10', '-m', 'src.train', training_type],
         stdout=subprocess.PIPE, 
         stderr=subprocess.PIPE
     )
-    
     subprocess.run(['python3.10', 'src/benchmarking/jetson_metrics.py', str(main_proc.pid)])
 
     stdout, stderr = main_proc.communicate()
@@ -15,4 +13,8 @@ def main():
     print(f'Main Errors: {stderr.decode()}')
 
 if __name__ == '__main__':
-    main()
+    if len(sys.argv) >= 2:
+        training_type = sys.argv[1]
+        measure(training_type)
+    else:
+        print(f'Usage: python3.x src.benchmarking.measure <training_type>')
