+++ src/q_learning/train_episodes.py
@@ -1,6 +1,6 @@
 import numpy as np
 from src.inverted_pendulum_simulator.src.inverted_pendulum import InvertedPendulum
-from src.train.q_learning import QLearning
+from src.q_learning.q_learning import QLearning
 
 
 def simulate_episodes(num_ep: int):
@@ -52,7 +52,6 @@ def simulate_episodes(num_ep: int):
             rewards_episode.append(reward)
             q_learning.update_q_table(disc_state, action, reward, next_state_disc, done)  # type: ignore
 
-        # print(f"Episode {episode_index} - Total reward: {sum(rewards_episode)}")
         total_rewards.append(sum(rewards_episode))
 
     return q_learning, total_rewards
